use namespaceWrapper fs

+++ arweave_task.js
@@ -1,33 +1,29 @@
 const Gatherer = require('./model/gatherer');
 const Arweave = require('./adapters/arweave/arweave');
-const { namespaceWrapper } = require('./namespaceWrapper');
+const {
+  namespaceWrapper,
+  taskNodeAdministered,
+} = require('./namespaceWrapper');
 const { Keypair } = require('@solana/web3.js'); // TEST For local testing only
 const fs = require('fs');
 const nacl = require('tweetnacl');
 const bs58 = require('bs58');
 const dataDb = require('./helpers/db');
-const fsPromise = require("fs/promises");
+const fsPromise = require('fs/promises');
 const { Web3Storage, getFilesFromPath } = require('web3.storage');
 const storageClient = new Web3Storage({
   token: process.env.SECRET_WEB3_STORAGE_KEY,
 });
 
-
 const credentials = {}; // arweave doesn't need credentials
 
-const run = async () => {
+const run = async round => {
   // Load node's keypair from the JSON file
   const keypair = await namespaceWrapper.getSubmitterAccount();
 
-let publicKeyHex = Buffer.from(keypair._keypair.publicKey).toString('hex');
+  let publicKeyHex = Buffer.from(keypair._keypair.publicKey).toString('hex');
 
   console.log('publicKeyHex', publicKeyHex);
-
-  // get Round
-  const round = await namespaceWrapper.getRound();
-
-  // TEST ROUND
-  // let round = 1000;
 
   // TEST For local testing, hardcode the keypair
   // const keypair = Keypair.generate();
@@ -66,37 +62,61 @@ let publicKeyHex = Buffer.from(keypair._keypair.publicKey).toString('hex');
   const proof_cid = await uploadIPFS(submission_value, round);
 
   return proof_cid;
-
 };
 
 async function uploadIPFS(data, round) {
   const proofPath = `./arweave/proofs${round}.json`;
 
-  if (!fs.existsSync(`./arweave`)) fs.mkdirSync(`./arweave`);
+  if (taskNodeAdministered) {
+    if (!namespaceWrapper.fs(access, `./arweave`)) {
+      namespaceWrapper.fs(mkdir, `./arweave`);
+    }
+  } else {
+    if (!fs.existsSync(`./arweave`)) fs.mkdirSync(`./arweave`);
+  }
 
   console.log('proofPATH', proofPath);
 
-  await fsPromise.writeFile(proofPath, JSON.stringify(data), (err) => {
-    if (err) {
-      console.error(err);
-    }
-  });
+  if (taskNodeAdministered) {
+    await namespaceWrapper.fs(writeFile, 
+      proofPath,
+      JSON.stringify(data),
+      err => {
+        if (err) {
+          console.error(err);
+        }
+      },
+    );
+  } else {
+    await fsPromise.writeFile(proofPath, JSON.stringify(data), err => {
+      if (err) {
+        console.error(err);
+      }
+    });
+  }
 
   if (storageClient) {
     const file = await getFilesFromPath(proofPath);
     const proof_cid = await storageClient.put(file);
     console.log(`Proofs of healthy list in round ${round} : `, proof_cid);
 
-    try {
-      await fsPromise.unlink(proofPath);
-    } catch (err) {
-      console.error(err);
+    if (taskNodeAdministered) {
+      try {
+        await namespaceWrapper.fs(unlink, proofPath);
+      } catch (err) {
+        console.error(err);
+      }
+    } else {
+      try {
+        await fsPromise.unlink(proofPath);
+      } catch (err) {
+        console.error(err);
+      }
     }
-    
     return proof_cid;
   } else {
     console.log('NODE DO NOT HAVE ACCESS TO WEB3.STORAGE');
   }
-};
+}
 
 module.exports = run;
